+++ services/minespace-web/src/components/Forms/projects/projectSummary/ProjectSummaryForm.tsx
@@ -16,7 +16,7 @@ import * as FORM from "@/constants/forms";
 import DocumentUpload from "@/components/Forms/projects/projectSummary/DocumentUpload";
 import ProjectContacts from "@/components/Forms/projects/projectSummary/ProjectContacts";
 import ProjectDates from "@/components/Forms/projects/projectSummary/ProjectDates";
-import AuthorizationsInvolved from "@/components/Forms/projects/projectSummary/AuthorizationsInvolved";
+import AuthorizationsInvolved from "@mds/common/components/projectSummary/AuthorizationsInvolved";
 import SteppedForm from "@mds/common/components/forms/SteppedForm";
 import Step from "@common/components/Step";
 import ProjectLinks from "@mds/common/components/projects/ProjectLinks";
